[Wave] Remove unused code in iree/turbine/transforms
