+++ iree/turbine/transforms/merger.py
@@ -4,16 +4,14 @@
 # See https://llvm.org/LICENSE.txt for license information.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-from typing import Any, Dict, List, Optional, Sequence, Union
+from typing import Dict, List, Optional, Sequence
 
 from iree.compiler.ir import (
     Attribute,
-    Block,
     InsertionPoint,
     Operation,
     StringAttr,
     SymbolTable,
-    Context,
 )
